make tracing optional

+++ coal/collection.go
@@ -20,10 +20,12 @@ type Collection struct {
 // AggregateAll wraps the native Aggregate collection method and decodes all
 // documents to the provided slice.
 func (c *Collection) AggregateAll(ctx context.Context, slicePtr interface{}, pipeline interface{}, opts ...*options.AggregateOptions) error {
-	// push span
-	c.trace.Push("coal/Collection.Aggregate")
-	c.trace.Tag("pipeline", pipeline)
-	defer c.trace.Pop()
+	// trace
+	if c.trace != nil {
+		c.trace.Push("coal/Collection.Aggregate")
+		c.trace.Tag("pipeline", pipeline)
+		defer c.trace.Pop()
+	}
 
 	// run query
 	csr, err := c.coll.Aggregate(ctx, pipeline, opts...)
@@ -44,10 +46,12 @@ func (c *Collection) AggregateAll(ctx context.Context, slicePtr interface{}, pip
 // provided callback with the decode method until an error is returned or the
 // cursor has been exhausted.
 func (c *Collection) AggregateIter(ctx context.Context, pipeline interface{}, fn func(func(interface{}) error) error, opts ...*options.AggregateOptions) error {
-	// push span
-	c.trace.Push("coal/Collection.Aggregate")
-	c.trace.Tag("pipeline", pipeline)
-	defer c.trace.Pop()
+	// trace
+	if c.trace != nil {
+		c.trace.Push("coal/Collection.Aggregate")
+		c.trace.Tag("pipeline", pipeline)
+		defer c.trace.Pop()
+	}
 
 	// run query
 	csr, err := c.coll.Aggregate(ctx, pipeline, opts...)
@@ -77,9 +81,12 @@ func (c *Collection) AggregateIter(ctx context.Context, pipeline interface{}, fn
 
 // BulkWrite wraps the native BulkWrite collection method.
 func (c *Collection) BulkWrite(ctx context.Context, models []mongo.WriteModel, opts ...*options.BulkWriteOptions) (*mongo.BulkWriteResult, error) {
-	// push span
-	c.trace.Push("coal/Collection.BulkWrite")
-	defer c.trace.Pop()
+	// trace
+	if c.trace != nil {
+		// push span
+		c.trace.Push("coal/Collection.BulkWrite")
+		defer c.trace.Pop()
+	}
 
 	// run query
 	return c.coll.BulkWrite(ctx, models, opts...)
@@ -87,10 +94,12 @@ func (c *Collection) BulkWrite(ctx context.Context, models []mongo.WriteModel, o
 
 // CountDocuments wraps the native CountDocuments collection method.
 func (c *Collection) CountDocuments(ctx context.Context, filter interface{}, opts ...*options.CountOptions) (int64, error) {
-	// push span
-	c.trace.Push("coal/Collection.CountDocuments")
-	c.trace.Log("filter", filter)
-	defer c.trace.Pop()
+	// trace
+	if c.trace != nil {
+		c.trace.Push("coal/Collection.CountDocuments")
+		c.trace.Log("filter", filter)
+		defer c.trace.Pop()
+	}
 
 	// run query
 	return c.coll.CountDocuments(ctx, filter, opts...)
@@ -98,10 +107,13 @@ func (c *Collection) CountDocuments(ctx context.Context, filter interface{}, opt
 
 // DeleteMany wraps the native DeleteMany collection method.
 func (c *Collection) DeleteMany(ctx context.Context, filter interface{}, opts ...*options.DeleteOptions) (*mongo.DeleteResult, error) {
-	// push span
-	c.trace.Push("coal/Collection.DeleteMany")
-	c.trace.Log("filter", filter)
-	defer c.trace.Pop()
+	// trace
+	if c.trace != nil {
+		// push span
+		c.trace.Push("coal/Collection.DeleteMany")
+		c.trace.Log("filter", filter)
+		defer c.trace.Pop()
+	}
 
 	// run query
 	return c.coll.DeleteMany(ctx, filter, opts...)
@@ -109,10 +121,13 @@ func (c *Collection) DeleteMany(ctx context.Context, filter interface{}, opts ..
 
 // DeleteOne wraps the native DeleteOne collection method.
 func (c *Collection) DeleteOne(ctx context.Context, filter interface{}, opts ...*options.DeleteOptions) (*mongo.DeleteResult, error) {
-	// push span
-	c.trace.Push("coal/Collection.DeleteOne")
-	c.trace.Log("filter", filter)
-	defer c.trace.Pop()
+	// trace
+	if c.trace != nil {
+		// push span
+		c.trace.Push("coal/Collection.DeleteOne")
+		c.trace.Log("filter", filter)
+		defer c.trace.Pop()
+	}
 
 	// run query
 	return c.coll.DeleteOne(ctx, filter, opts...)
@@ -120,11 +135,14 @@ func (c *Collection) DeleteOne(ctx context.Context, filter interface{}, opts ...
 
 // Distinct wraps the native Distinct collection method.
 func (c *Collection) Distinct(ctx context.Context, fieldName string, filter interface{}, opts ...*options.DistinctOptions) ([]interface{}, error) {
-	// push span
-	c.trace.Push("coal/Collection.Distinct")
-	c.trace.Tag("fieldName", fieldName)
-	c.trace.Log("filter", filter)
-	defer c.trace.Pop()
+	// trace
+	if c.trace != nil {
+		// push span
+		c.trace.Push("coal/Collection.Distinct")
+		c.trace.Tag("fieldName", fieldName)
+		c.trace.Log("filter", filter)
+		defer c.trace.Pop()
+	}
 
 	// run query
 	return c.coll.Distinct(ctx, fieldName, filter, opts...)
@@ -132,9 +150,12 @@ func (c *Collection) Distinct(ctx context.Context, fieldName string, filter inte
 
 // EstimatedDocumentCount wraps the native EstimatedDocumentCount collection method.
 func (c *Collection) EstimatedDocumentCount(ctx context.Context, opts ...*options.EstimatedDocumentCountOptions) (int64, error) {
-	// push span
-	c.trace.Push("coal/Collection.EstimatedDocumentCount")
-	defer c.trace.Pop()
+	// trace
+	if c.trace != nil {
+		// push span
+		c.trace.Push("coal/Collection.EstimatedDocumentCount")
+		defer c.trace.Pop()
+	}
 
 	// run query
 	return c.coll.EstimatedDocumentCount(ctx, opts...)
@@ -143,10 +164,13 @@ func (c *Collection) EstimatedDocumentCount(ctx context.Context, opts ...*option
 // FindAll wraps the native Find collection method and decodes all documents to
 // the provided slice.
 func (c *Collection) FindAll(ctx context.Context, slicePtr interface{}, filter interface{}, opts ...*options.FindOptions) error {
-	// push span
-	c.trace.Push("coal/Collection.Find")
-	c.trace.Tag("filter", filter)
-	defer c.trace.Pop()
+	// trace
+	if c.trace != nil {
+		// push span
+		c.trace.Push("coal/Collection.Find")
+		c.trace.Tag("filter", filter)
+		defer c.trace.Pop()
+	}
 
 	// run query
 	csr, err := c.coll.Find(ctx, filter, opts...)
@@ -167,10 +191,13 @@ func (c *Collection) FindAll(ctx context.Context, slicePtr interface{}, filter i
 // callback with the decode method until an error is returned or the cursor has
 // been exhausted.
 func (c *Collection) FindIter(ctx context.Context, filter interface{}, fn func(func(interface{}) error) error, opts ...*options.FindOptions) error {
-	// push span
-	c.trace.Push("coal/Collection.Find")
-	c.trace.Tag("filter", filter)
-	defer c.trace.Pop()
+	// trace
+	if c.trace != nil {
+		// push span
+		c.trace.Push("coal/Collection.Find")
+		c.trace.Tag("filter", filter)
+		defer c.trace.Pop()
+	}
 
 	// run query
 	csr, err := c.coll.Find(ctx, filter, opts...)
@@ -200,10 +227,13 @@ func (c *Collection) FindIter(ctx context.Context, filter interface{}, fn func(f
 
 // FindOne wraps the native FindOne collection method.
 func (c *Collection) FindOne(ctx context.Context, filter interface{}, opts ...*options.FindOneOptions) lungo.ISingleResult {
-	// push span
-	c.trace.Push("coal/Collection.FindOne")
-	c.trace.Log("filter", filter)
-	defer c.trace.Pop()
+	// trace
+	if c.trace != nil {
+		// push span
+		c.trace.Push("coal/Collection.FindOne")
+		c.trace.Log("filter", filter)
+		defer c.trace.Pop()
+	}
 
 	// run query
 	return c.coll.FindOne(ctx, filter, opts...)
@@ -211,10 +241,13 @@ func (c *Collection) FindOne(ctx context.Context, filter interface{}, opts ...*o
 
 // FindOneAndDelete wraps the native FindOneAndDelete collection method.
 func (c *Collection) FindOneAndDelete(ctx context.Context, filter interface{}, opts ...*options.FindOneAndDeleteOptions) lungo.ISingleResult {
-	// push span
-	c.trace.Push("coal/Collection.FindOneAndDelete")
-	c.trace.Log("filter", filter)
-	defer c.trace.Pop()
+	// trace
+	if c.trace != nil {
+		// push span
+		c.trace.Push("coal/Collection.FindOneAndDelete")
+		c.trace.Log("filter", filter)
+		defer c.trace.Pop()
+	}
 
 	// run query
 	return c.coll.FindOneAndDelete(ctx, filter, opts...)
@@ -222,10 +255,13 @@ func (c *Collection) FindOneAndDelete(ctx context.Context, filter interface{}, o
 
 // FindOneAndReplace wraps the native FindOneAndReplace collection method.
 func (c *Collection) FindOneAndReplace(ctx context.Context, filter interface{}, replacement interface{}, opts ...*options.FindOneAndReplaceOptions) lungo.ISingleResult {
-	// push span
-	c.trace.Push("coal/Collection.FindOneAndReplace")
-	c.trace.Log("filter", filter)
-	defer c.trace.Pop()
+	// trace
+	if c.trace != nil {
+		// push span
+		c.trace.Push("coal/Collection.FindOneAndReplace")
+		c.trace.Log("filter", filter)
+		defer c.trace.Pop()
+	}
 
 	// run query
 	return c.coll.FindOneAndReplace(ctx, filter, replacement, opts...)
@@ -233,10 +269,13 @@ func (c *Collection) FindOneAndReplace(ctx context.Context, filter interface{}, 
 
 // FindOneAndUpdate wraps the native FindOneAndUpdate collection method.
 func (c *Collection) FindOneAndUpdate(ctx context.Context, filter interface{}, update interface{}, opts ...*options.FindOneAndUpdateOptions) lungo.ISingleResult {
-	// push span
-	c.trace.Push("coal/Collection.FindOneAndUpdate")
-	c.trace.Log("filter", filter)
-	defer c.trace.Pop()
+	// trace
+	if c.trace != nil {
+		// push span
+		c.trace.Push("coal/Collection.FindOneAndUpdate")
+		c.trace.Log("filter", filter)
+		defer c.trace.Pop()
+	}
 
 	// run query
 	return c.coll.FindOneAndUpdate(ctx, filter, update, opts...)
@@ -244,9 +283,12 @@ func (c *Collection) FindOneAndUpdate(ctx context.Context, filter interface{}, u
 
 // InsertMany wraps the native InsertMany collection method.
 func (c *Collection) InsertMany(ctx context.Context, documents []interface{}, opts ...*options.InsertManyOptions) (*mongo.InsertManyResult, error) {
-	// push span
-	c.trace.Push("coal/Collection.InsertMany")
-	defer c.trace.Pop()
+	// trace
+	if c.trace != nil {
+		// push span
+		c.trace.Push("coal/Collection.InsertMany")
+		defer c.trace.Pop()
+	}
 
 	// run query
 	return c.coll.InsertMany(ctx, documents, opts...)
@@ -254,9 +296,12 @@ func (c *Collection) InsertMany(ctx context.Context, documents []interface{}, op
 
 // InsertOne wraps the native InsertOne collection method.
 func (c *Collection) InsertOne(ctx context.Context, document interface{}, opts ...*options.InsertOneOptions) (*mongo.InsertOneResult, error) {
-	// push span
-	c.trace.Push("coal/Collection.InsertOne")
-	defer c.trace.Pop()
+	// trace
+	if c.trace != nil {
+		// push span
+		c.trace.Push("coal/Collection.InsertOne")
+		defer c.trace.Pop()
+	}
 
 	// run query
 	return c.coll.InsertOne(ctx, document, opts...)
@@ -264,10 +309,13 @@ func (c *Collection) InsertOne(ctx context.Context, document interface{}, opts .
 
 // ReplaceOne wraps the native ReplaceOne collection method.
 func (c *Collection) ReplaceOne(ctx context.Context, filter interface{}, replacement interface{}, opts ...*options.ReplaceOptions) (*mongo.UpdateResult, error) {
-	// push span
-	c.trace.Push("coal/Collection.ReplaceOne")
-	c.trace.Log("filter", filter)
-	defer c.trace.Pop()
+	// trace
+	if c.trace != nil {
+		// push span
+		c.trace.Push("coal/Collection.ReplaceOne")
+		c.trace.Log("filter", filter)
+		defer c.trace.Pop()
+	}
 
 	// run query
 	return c.coll.ReplaceOne(ctx, filter, replacement, opts...)
@@ -275,10 +323,13 @@ func (c *Collection) ReplaceOne(ctx context.Context, filter interface{}, replace
 
 // UpdateMany wraps the native UpdateMany collection method.
 func (c *Collection) UpdateMany(ctx context.Context, filter interface{}, update interface{}, opts ...*options.UpdateOptions) (*mongo.UpdateResult, error) {
-	// push span
-	c.trace.Push("coal/Collection.UpdateMany")
-	c.trace.Log("filter", filter)
-	defer c.trace.Pop()
+	// trace
+	if c.trace != nil {
+		// push span
+		c.trace.Push("coal/Collection.UpdateMany")
+		c.trace.Log("filter", filter)
+		defer c.trace.Pop()
+	}
 
 	// run query
 	return c.coll.UpdateMany(ctx, filter, update, opts...)
@@ -286,10 +337,13 @@ func (c *Collection) UpdateMany(ctx context.Context, filter interface{}, update 
 
 // UpdateOne wraps the native UpdateOne collection method.
 func (c *Collection) UpdateOne(ctx context.Context, filter interface{}, update interface{}, opts ...*options.UpdateOptions) (*mongo.UpdateResult, error) {
-	// push span
-	c.trace.Push("coal/Collection.UpdateOne")
-	c.trace.Log("filter", filter)
-	defer c.trace.Pop()
+	// trace
+	if c.trace != nil {
+		// push span
+		c.trace.Push("coal/Collection.UpdateOne")
+		c.trace.Log("filter", filter)
+		defer c.trace.Pop()
+	}
 
 	// run query
 	return c.coll.UpdateOne(ctx, filter, update, opts...)
